Move the parsers into one place
