+++ src/comparison/compare.ts
@@ -2,8 +2,8 @@ import type { Change } from "diff"
 import isEqual from "lodash.isequal"
 import CoreHandler from "src/index"
 import convertAhoToXml from "src/lib/generateLegacyAhoXml"
-import parseAhoXml from "src/lib/parseAhoXml"
 import { xmlOutputDiff, xmlOutputMatches } from "src/lib/xmlOutputComparison"
+import parseAhoXml from "src/parse/parseAhoXml"
 import type Exception from "src/types/Exception"
 import type { Trigger } from "src/types/Trigger"
 import extractExceptionsFromAho from "tests/helpers/extractExceptionsFromAho"
