+++ src/index.ts
@@ -1,12 +1,12 @@
-import parseAhoXml from "./lib/parseAhoXml"
+import parseAhoXml from "./parse/parseAhoXml"
+import parseSpiResult from "./parse/parseSpiResult"
+import transformSpiToAho from "./parse/transformSpiToAho"
 import type { AnnotatedHearingOutcome } from "./types/AnnotatedHearingOutcome"
 import type BichardResultType from "./types/BichardResultType"
 import type PncGateway from "./types/PncGateway"
 import enrichHearingOutcome from "./use-cases/enrichHearingOutcome"
 import generateExceptions from "./use-cases/generateExceptions"
 import generateTriggers from "./use-cases/generateTriggers"
-import parseSpiResult from "./use-cases/parseSpiResult"
-import transformSpiToAho from "./use-cases/transformSpiToAho"
 
 export default (message: string, pncGateway: PncGateway): BichardResultType => {
   let hearingOutcome: AnnotatedHearingOutcome | Error
