+++ src/parse/transformSpiToAho/lookupAmountTypeByCjsCode.ts
