+++ src/use-cases/enrichHearingOutcome/enrichFunctions/enrichCourt.test.ts
@@ -1,6 +1,6 @@
 import { readFileSync } from "fs"
-import parseSpiResult from "src/use-cases/parseSpiResult"
-import transformSpiToAho from "src/use-cases/transformSpiToAho"
+import parseSpiResult from "src/parse/parseSpiResult"
+import transformSpiToAho from "src/parse/transformSpiToAho"
 import enrichCourt from "./enrichCourt"
 
 const message = readFileSync("test-data/input-message-001.xml", "utf-8")
