+++ src/use-cases/enrichHearingOutcome/enrichFunctions/enrichWithPncQuery.test.ts
@@ -1,7 +1,7 @@
+import parseSpiResult from "src/parse/parseSpiResult"
+import transformSpiToAho from "src/parse/transformSpiToAho"
 import type { AnnotatedHearingOutcome } from "src/types/AnnotatedHearingOutcome"
 import type PncGateway from "src/types/PncGateway"
-import parseSpiResult from "src/use-cases/parseSpiResult"
-import transformSpiToAho from "src/use-cases/transformSpiToAho"
 import generateMessage from "tests/helpers/generateMessage"
 import generateMockPncQueryResult from "tests/helpers/generateMockPncQueryResult"
 import MockPncGateway from "tests/helpers/MockPncGateway"
