+++ tests/comparison.test.ts
@@ -3,7 +3,7 @@ import "jest-xml-matcher"
 import orderBy from "lodash.orderby"
 import CoreHandler from "src/index"
 import convertAhoToXml from "src/lib/generateLegacyAhoXml"
-import parseAhoXml from "src/lib/parseAhoXml"
+import parseAhoXml from "src/parse/parseAhoXml"
 import type Exception from "src/types/Exception"
 import extractExceptionsFromAho from "./helpers/extractExceptionsFromAho"
 import generateMockPncQueryResultFromAho from "./helpers/generateMockPncQueryResultFromAho"
