+++ tests/helpers/mockRecordInPnc.ts
@@ -1,7 +1,7 @@
 import axios from "axios"
 import merge from "lodash.merge"
+import parseSpiResult from "src/parse/parseSpiResult"
 import type { OffenceParsedXml, ResultedCaseMessageParsedXml } from "src/types/IncomingMessage"
-import parseSpiResult from "src/use-cases/parseSpiResult"
 import defaults from "./defaults"
 import reformatDate from "./reformatDate"
